Move files to java directory
