+++ app/src/main/java/de/markusfisch/android/shadereditor/highlighter/JniLexer.java
