+++ app/src/main/java/de/markusfisch/android/shadereditor/highlighter/Trie.java
